refactored components and set up App.js
