+++ lost-and-found-react/package.json
@@ -3,8 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome": "^1.1.8",
+    "@fortawesome/fontawesome-free-brands": "^5.0.13",
+    "@fortawesome/fontawesome-svg-core": "^1.2.12",
+    "@fortawesome/free-solid-svg-icons": "^5.6.3",
+    "@fortawesome/react-fontawesome": "^0.1.3",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
+    "react-notify-toast": "^0.5.0",
     "react-scripts": "2.1.3"
   },
   "scripts": {
