+++ lost-and-found-react/src/App.css
@@ -1,3 +1,15 @@
+.header{
+  display: flex;
+  justify-content: space-around;
+}
+.header > h1{
+  width: 50%;
+}
+
+.actionButtons{
+  margin: auto;
+}
+
 body {
   font-family: Roboto, sans-serif;
   background: #f2f2f2;
