+++ lost-and-found-react/src/components/Item.js
@@ -1,40 +1,40 @@
 
-import React, {Component} from "react";
+// import React, {Component} from "react";
 
 
-class Items extends Component{
-    renderItems(){
-        return this.props.items.map((item, index) => {
-                return (
-                    <div className="" key={index}>
-                        <div>
-                            <img src={this.props.items.imageUrl} alt="" />
-                        </div>
-                        <div>
-                            <p>Name:{item.name}</p>
-                          <select>
-                            <option value="found"></option>
-                            <option value="lost"></option>
-                            {/* <p>Type:{item.type}</p> */}
-                            </select>
+// class Items extends Component{
+//     renderItems(){
+//         return this.props.items.map((item, index) => {
+//                 return (
+//                     <div className="" key={index}>
+//                         <div>
+//                             <img src={this.props.items.imageUrl} alt="" />
+//                         </div>
+//                         <div>
+//                             <p>Name:{item.name}</p>
+//                           <select>
+//                             <option value="found"></option>
+//                             <option value="lost"></option>
+//                             {/* <p>Type:{item.type}</p> */}
+//                             </select>
                             
-                            <p>Added By:{item.addedBy}</p>
-                            <p>Added date{items.addedDate}</p>
-                        </div>
+//                             <p>Added By:{item.addedBy}</p>
+//                             <p>Added date{items.addedDate}</p>
+//                         </div>
 
 
-                    </div>
-                )
-            })
-        }
+//                     </div>
+//                 )
+//             })
+//         }
 
 
-   render() {
-       return(<div> {this.renderItems() }</div>);
-   }
-}
+//    render() {
+//        return(<div> {this.renderItems() }</div>);
+//    }
+// }
 
 
 
 
-export default Items;
+// export default Items;
