+++ lost-and-found-react/src/components/imageUpload/Footer.js
